fix: address maintainer review - forceCreate bypass, null client early return, extract embedding utility, move config to semantic-cache.config.ts
